third_party/WebKit/LayoutTests/css3/calc/webkit-line-clamp-calc.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: calc() function in webkit-line-clamp</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
    #test {
        -webkit-line-clamp: calc(2);
    }
</style>
<div id="test"></div>
<script>
    const div = document.querySelector("#test");
    test(function() {
        assert_equals(getComputedStyle(div).webkitLineClamp, "2");
    }, "calc() in webkit-line-clamp");
</script>
